+++ core/src/main/java/org/neo4j/graphalgo/core/utils/paged/PagedLongDoubleMap.java
@@ -34,7 +34,7 @@ public final class PagedLongDoubleMap {
 
     private static final int PAGE_SHIFT = 14;
     static final int PAGE_SIZE = 1 << PAGE_SHIFT;
-    private static final long PAGE_MASK = (long) (PAGE_SIZE - 1);
+    private static final long PAGE_MASK = PAGE_SIZE - 1;
 
     private static final MemoryEstimation MEMORY_REQUIREMENTS = MemoryEstimations
             .builder(PagedLongDoubleMap.class)
